[KNI][CARRY] noderesourcetopology: return informer in init
In order to make room for the upcoming Reserve plugin support,
we change the *internal* API of the initialization helper, to

return the initialized Informer - not the Lister straight away.
The orchestrating function will do the high level wiring, like
creating and passing around the Lister.

No intended change in behaviour, no external API changes.

Signed-off-by: Francesco Romani <fromani@redhat.com>