[KNI][CARRY] noderesourcetopology: add getNodeTopology helper
In order to support the upcoming Reserve plugin, we
need an abstraction for the Filter and Score plugins
to get the adjusted topology data.

We make room for this by introducing the getNodeToplogy helper;
this way the Filter and Score plugins will require no further
changes while the Reserve implementation will be fleshed out.

Signed-off-by: Francesco Romani <fromani@redhat.com>