[KNI][CARRY] noderesourcetopology: track pending resources
Up until now, the `filter` and `score` plugin depend on the
freshness of the noderesourcetopology (NRT) objects to make their
decisions. IOW, the plugins depend on timely reporting of NRT objects
to learn about the actual resource availability; they consider
the resources free until told otherwise. This leads to resource
overallocation, which in turn increases the likeness of
TopologyAffinityError, which is what the plugins wants to avoid
(or at least reduce as much as possible) in the first place.

We call this behaviour `false positive`: the scheduler plugin wrongly
believes resources are available, while they are not, and schedule
a pod while it should wait (re-enqueue for later).
Similarly, we then define `false negative` the opposite behaviour,
on which the scheduler plugin keeps the pod unscheduled (Pending)
because it wrongly believes resources are NOT available.

This change introduces a NRT-aware `reserve` plugin, which tracks the
pending resources, which the scheduler knows have been allocated to
pods, but which aren't yet been reported as un-available by the
NRT updater.

The `reserve` plugin is optional, simple as possible, requires minimal
changes in the `filter` and `score` plugin, tolerates restarts well,
and is backward compatible, meaning that `score` and `filter` should see
NO CHANGES in their behaviour with the `reserve` plugin disabled.

On the other hand, with `reserve` enabled, the resources allocated to
pods are decremented, hence the `filter` and `score` see a more
realistic, albeit more pessimist (out of necessity) picture.

This has two main benefits:
1. we have now the option to greatly reduce the pressure on the topology
   updater, because good scheduling decision is less dependent on timely
   update of NRT objects (`reserve` of course reconciles its own state
   with fresh NRT state once available)
2. we add a knob to cluster administrators to prefer to keep pods in
   waiting state (false negatives using the the definition above) and
   have less frequent topology updates.

Overall, the `reserve` plugin allows to optionally reduce the pressure
on the topology update frequency while being fully backward compatible
and optional.

Signed-off-by: Francesco Romani <fromani@redhat.com>