[KNI] introducing KNI-specific plugin
A very ambitious but important goal of this fork (in github sense) is
to keep full compatibility with the upstream project, and to have
very obvious, crystal clear paper trail about changes.
Ideally, we would like to have the very same u/s commits (SHA-level
identical) and to add on top of that the bits and pieces we need.

At the same time, another key goal is to carry patches u/s is not (yet)
ready to accept, to experiment and/or to full fill KNI specific needs.

These two goals conflict to a significant extent. Carry u/s patches
out of order (or which u/s didn't accept) alter the git history,
and this makes harder or impossible to keep 1:1 relationship
between u/s commits and KNI-specific commits.

It's possible that in the future we will need to relax this constraint,
and to deal the fact that we cannot any longer just have SHA-level
identity between the histories of these two repos.

But we need to try hard anyway.
A possible way out is to keep an interna clone of the
noderesourcetopology package, to hold here the carried patches.
Over time, u/s (hopefully) accepts these changes and/or will provide
better replacements; in these cases we will resync or reset the
staging area periodically to realign with the reference sources,
to the point that the two codebase are identical and aliasing each
other.

Signed-off-by: Francesco Romani <fromani@redhat.com>