test: polish integration tests to use envtest
This change removes the dependency on
apiservertesting testing package and leverage
the envtest framework from controller-runtime
https://github.com/kubernetes-sigs/controller-runtime/tree/master/pkg/envtest
to spin up apiserver and etcd in integration test

Using this framework will simplify the code
and remove the need for hack/testdata and
the autogen target that uses update-generated-openapi.sh
The bring up of apiserver and etcd is done in the
TestMain so that we only need to do this once and
the integration tests will not need to spin up
apiserver each time which was the case before.

As part of the this rework, ensured that all pods
are properly cleaned up and removed at the end
of each integration test.

Signed-off By: Yibo Zhuang <yibzhuang@gmail.com>

(cherry picked from commit 8ddf60bb590314da62a68dabc5db208cfe6f3ba1)