implement strict alignment check for the noderesourcetopology filter plugin (#347)
* nrt: filter: use IsZero()

Use resource.Quantity.IsZero() helper instead of comparing
with zero quantity. This is a bit nicer to read, but most notably
saves us a relatively costly MustParse call in each function run.

No intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: isolate the per-NUMA decision logic

Extract the inner logic to check if a NUMA Node is suitable
in its own function.
This change enables further cleanups, and easier changing
of the filtering logic.
It will also make adding more logs easier down the line

No intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: split complex if in a sequence of ifs

Split the complex condition, to make the code easier
to read, change, and possibly add (and log) the
NUMA node reject reason in the future.

No intended change in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: qos=guaranteed requires align for mem/hp

In kubernetes >= 1.23, which is already out, we can
depend on the memory manager for alignment of memory and
hugepages. So we make the alignment check consistent to
what we do for CPU.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: make test more extendible

Make it easier to add new test cases without intended
changes in behaviour

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: flter: add more test cases

Improve the test coverage of the filter plugin

Signed-off-by: Francesco Romani <fromani@redhat.com>