introduce the KNIDebug plugin
The KNIDebug is meant to be a passthrough plugin (which will
fail only in the most catastrophic of cases) which we can use
to inject extra logs u/s is not willing or interesting to accept.

Using the KNIDebug plugin, we can choose between the tradeoff
of changing the u/s code (which increases the maintainership
costs) at the cost of limiting ourselves in the logs we can add:
some logs can only meaningfully by added in the private function
of some plugins.

Signed-off-by: Francesco Romani <fromani@redhat.com>