nrt: fix filter plugin resource alignment when scope=container
The current logic examines each container individually without taking into account the resources that are allegedly already allocated for previous containers.

In order to fix that, for every container, we should subtruct the resources requested by the container from the available resources, before checking the next container.
This way we can guaranteed an accurate snapshot of the available resource.

The problem with this technique is that we need to predict Kubelet's behavior and knowing in advance from which NUMA ID the container's resources are about to get allocated.
Careful examination of Kubelet flow, when scope=container and policy=single-numa-node shows that:
- Kubelet consider the NUMA zone as a possibility, only if all the requested CPUs by the container can be allocated from it.
  https://github.com/kubernetes/kubernetes/blob/v1.24.0-rc.1/pkg/kubelet/cm/topologymanager/policy.go#L155
- If Kubelet has multiple NUMAs selection from the above step, it will prefer the one which has the narrowest mask.
  A mask is said to be "narrower" than another if it has less bits set. If the same number of bits are set in both masks, then the mask with more lower-numbered bits set wins out.
  https://github.com/kubernetes/kubernetes/blob/v1.24.0-rc.1/pkg/kubelet/cm/topologymanager/bitmask/bitmask.go#L141

The filter plugin is using this information in order to predict Kubelet’s behavior and subtracts the available resources from the correct NUMA zone while doing the filtering process when scope=container.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>