nrt: integration-tests: adding scope=container tests
Add batch of tests that are validating the flow of NodeResourceTopology plugin when Kubelet configured with TopologyManagerScope=container.

In case of tests that are expecting the pod to keep in Pending state, we are expecting
an Event with a very specific error message that should be reported by the NodeResourceTopology plugin.
This is done in order to make sure the test is not falling due to
other plugins like NodeResourceFit plugin.

Each test has a type and tier level which are correspond to the table in the TESTS.md file.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>
(cherry picked from commit b86cfe93d9d678d5e7f3b0b36eacbeba2173be24)