nrt: document tests
When the plugin is running with scope=container configured, there are many possible scenarios that needed a coverage by the tests.
In order to track the coverage more easly,let's have a document describing all the possible cases and permutations we can have.

In the tests themselves we'll refer to the relevant entry on the table in the document.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>