[KNI] log: add the logtracr logger
Introduce the logtracr logger.
The logtracr logger builds upon and augments the existing klog (+logr)
infrastructure to provide per-object (vs per-component) log.
This logger implement a simplistic, in-process lookalike of a tracing
logger, whose concepts (spans and traces) kinda remind of
opentelemetry's concepts.

In a nutshell, the logger aims to make it possible/easier to correlate
all the logs pertaining to a container during a scheduling cycle

For full design details please check the upcoming
`pkg-kni/logtracr/DESIGN.md`.

Signed-off-by: Francesco Romani <fromani@redhat.com>