[KNI] go.mod: bump x/oauth2 0.23.0->0.30.0
To fix cve-2025-22868, we were advised internally to use a forked
`x/oauth2` by openshift consdiring that the fix is consuming v0.27.0+ of
x/oauth2 but that needed Go v1.23.
Now that the project is using 1.23 Go modules, bump the library to
v0.30.0 (latest so far)  and run `make -f Makefile.kni update-vendor`.
ref: https://github.com/golang/go/issues/71490

Signed-off-by: Shereen Haj <shajmakh@redhat.com>