nrt: score: introduce and use scoreInfo
similarly to what we did for filter, add a helper struct
to group together all the context we need to carry in the
internal implementation functions.
The goals are to streamline the existing code, to
reduce redundant call and to make room for more
arguments to be passed through, like the topology manager
configuration.

Signed-off-by: Francesco Romani <fromani@redhat.com>