WIP: [KNI] pfpstatus: optional HTTP endpoint
When exposing pfpstatus data, dumping to files was
fallback because we couldn't find a quick way to expose
the data using HTTP in a secure way.

Things are  slowly changing, so this PR explores the option
to add a HTTP endpoint to inspect the pfpstatus.

The internal collection must be extremely cheap, but the
publication of data doesn't need to be so cheap (but still
must be as lightweight as possible), because
is done very sporadically when troubleshooting, possibly
just once in a while using must-gather.

The only main remaining blocker is how to secure
the new endpoint using kube/OCP primitives.

AI: used for the authMiddleware bootstrap

AI-Attribution: AIA Ph Nc Hin R gemini-2.5-pro v1.0
Signed-off-by: Francesco Romani <fromani@redhat.com>